Move controller test files into controller package.
These test files clearly test packages in org.tvrenamer.controller so that's where they should be, too.  This commit doesn't modify any actual tests, just moves the files and updates the package and import lines.
