+++ src/test/org/tvrenamer/controller/TVRenamerTest.java
@@ -1,4 +1,4 @@
-package org.tvrenamer;
+package org.tvrenamer.controller;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -7,8 +7,6 @@ import static org.junit.Assert.fail;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import org.tvrenamer.controller.ShowInformationListener;
-import org.tvrenamer.controller.TVRenamer;
 import org.tvrenamer.model.FileEpisode;
 import org.tvrenamer.model.Show;
 import org.tvrenamer.model.ShowStore;
